+++ app/scripts/fhir/R4/lformsFHIR.js
@@ -19225,7 +19225,7 @@ var sdcExport = {
       this._handleExternallyDefined(targetItem, item);
     } // option, for answer list
     else if (item.answers) {
-        targetItem.option = this._handleAnswers(item, noExtensions);
+        targetItem.answerOption = this._handleAnswers(item, noExtensions);
       } // initialValue, for default values
 
 
@@ -19655,10 +19655,14 @@ var sdcExport = {
 
 
       option.valueCoding = {
-        "system": "http://loinc.org",
         "code": answer.code,
         "display": answer.text
       };
+
+      if (item.answerCodeSystem) {
+        option.valueCoding.system = this._getCodeSystem(item.answerCodeSystem);
+      }
+
       optionArray.push(option);
     }
 
@@ -20575,25 +20579,29 @@ function addSDCImportFns(ns) {
 
 
   function _processAnswers(lfItem, qItem) {
-    if (qItem.option) {
+    if (qItem.answerOption) {
       lfItem.answers = [];
 
-      for (var i = 0; i < qItem.option.length; i++) {
+      for (var i = 0; i < qItem.answerOption.length; i++) {
         var answer = {};
-        var label = LForms.Util.findObjectInArray(qItem.option[i].extension, 'url', self.fhirExtUrlOptionPrefix);
+        var label = LForms.Util.findObjectInArray(qItem.answerOption[i].extension, 'url', self.fhirExtUrlOptionPrefix);
 
         if (label) {
           answer.label = label.valueString;
         }
 
-        var score = LForms.Util.findObjectInArray(qItem.option[i].modifierExtension, 'url', self.fhirExtUrlOptionScore);
+        var score = LForms.Util.findObjectInArray(qItem.answerOption[i].modifierExtension, 'url', self.fhirExtUrlOptionScore);
 
         if (score) {
           answer.score = score.valueInteger.toString();
         }
 
-        answer.code = qItem.option[i].valueCoding.code;
-        answer.text = qItem.option[i].valueCoding.display;
+        if (qItem.answerOption[i].valueCoding.system) {
+          qItem.answerCodeSystem = qItem.answerOption[i].valueCoding.system;
+        }
+
+        answer.code = qItem.answerOption[i].valueCoding.code;
+        answer.text = qItem.answerOption[i].valueCoding.display;
         lfItem.answers.push(answer);
       }
     }